Reorganize query resources into individual files
